==> activity10/01-csv-01.ipynb <==
# Reading a CSV as a Pandas DataFrame 

### 1. Import libraries and dependencies

# Import pandas and pathlib
import pandas as pd
from pathlib import Path

### 2. Create a Path to the File Using Pathlib

# Use the Pathlib libary to set the path to the CSV
csvpath = Path("./Resources/people.csv")

### 3. Read the CSV into a Pandas DataFrame

# Use the file path to read the CSV into a DataFrame
people_df = pd.read_csv(csvpath)

### 4. Display a Few Rows of the DataFrame

# Use the `head` function to display a few rows of the top of the DataFrame
people_df.head()

### 5. Generate Summary Statistics on the DataFrame

# Use the `include` parameter set to 'all' to include all columns (and datatypes)
people_df.describe(include='all')
