==> activity10/02_columns-01.ipynb <==
# Column Manipulation of Pandas DataFrames

### 1. Import libraries and dependencies

# Import pandas, pathlib, and numpy libraries
import pandas as pd
from pathlib import Path
import numpy as np

### 2. Create a Path to the File Using Pathlib

# Use the Pathlib libary to set the path to the CSV
csvpath = Path("./Resources/people.csv")

### 3. Read the CSV into a Pandas DataFrame and Display a Few Rows

# Use the file path to read the CSV into a DataFrame and display a few rows
people_df = pd.read_csv(csvpath)
people_df.head()

### 4. View Columns Names

# Use the `columns` attribute to output the column names
people_df.columns

### 5. View Column Data Types

# Use the `dtypes` attribute to output the column names and data types
people_df.dtypes

### 6. Rename Columns Using List

# Set the `columns` attribute to a new list of column names
column_names = ["Person_ID", "First_Name", "Last_Name", "Email", "Gender", "University", "Occupation", "Salary"]
people_df.columns = column_names
people_df.head()

### 7. Rename Columns Using Dictionary

# Use the `rename` function and set the `columns` parameter to a dictionary of new column names
people_df = people_df.rename(columns={
    "id" : "Person_ID",
    "first_name" : "First_Name",
    "last_name" : "Last_Name",
    "email" : "Email",
    "gender" : "Gender",
    "uni_grad" : "University",
    "job_title" : "Occupation",
    "Income" : "Salary"

    
})

people_df.head()

### 8. Re-order Columns

# Use a list of re-ordered column names to alter the column order of the original DataFrame
people_df = people_df[["Person_ID", "Last_Name", "First_Name", "Gender", "University", "Occupation", "Salary", "Email"]]
                      
people_df.head()                                               

### 9. Create Additional Column

# Use the `randint` function to randomly generate an `Age` from 22 to 65 for 1000 rows
people_df["Age"] = np.random.randint(low=22, high=65, size=1000)
people_df["Age_Copy"] = np.random.randint(low=22, high=65, size=1000)
people_df.head()

### 10. Delete Columns

# Use the `drop` function to delete the newly created `Age` column
people_df = people_df.drop(columns=["Age_Copy"])
people_df.head()

### 11. Save Modified DataFrame to New CSV

# Save the DataFrame to the `Resources` folder
people_df.to_csv("./Resources/people_reordered.csv")
