==> activity10/06_groupby-01.ipynb <==
# Grouping Data in Pandas DataFrames

## 1. Import Libraries and Dependencies

# Import necessary libraries and dependencies
import pandas as pd
from pathlib import Path

## 2. Create a Path to the File Using Pathlib

# Use the Pathlib libary to set the path to the CSV
csvpath = Path("./Resources/people_cleansed.csv") 

## 3. Read the CSV into a Pandas DataFrame

# Use the file path to read the CSV into a DataFrame and display a few rows
people_cleansed_df = pd.read_csv(csvpath)
people_cleansed_df.head()

## 4. Group DataFrame by `Occupation` and perform `count` aggregation

# Group by `Occupation` and perform count
people_cleansed_grp = people_cleansed_df.groupby('Occupation').count()
people_cleansed_grp

## 5. Group DataFrame by `Occupation` and Calculate Average Salary and Age

# Calculate average Salary and Age for each Occupation
people_cleansed_df.groupby('Occupation').mean()

## 6. Group By `Occupation` and `Gender` Columns, then Calculate Average Salary and Age

# Group by Occupation and Gender columns
people_cleansed_df.groupby(['Occupation', 'Gender']).mean()
