==> activity10/07_binning-01.ipynb <==
# Categorizing Records in Pandas DataFrames

## 1. Import Libraries and Dependencies

# import dependencies
import pandas as pd
from pathlib import Path

## 2. Set the File Path to the CSV using Pathlib

# Use the Pathlib libary to set the path to the CSV
csvpath = Path("./Resources/people_cleansed.csv")

## 3. Import the CSV into Pandas DataFrame

# Use the file path to read the CSV into a DataFrame and display a few rows
people_df = pd.read_csv(csvpath)
people_df.head()

## 4. Create Bins for Total Ranking

# Create the variable bins to define our bounds for when we cut the dataframe.
bins = [0, 30000, 70000, 100000, 200000]

# Create names for the bins
bin_names = ['Low', 'Moderate', 'Above Average', 'High']

## 5. Label Records According to the Corresponding Bin

# Add a column named "Salary Level" to people_df and slice the data into the bins
people_df["Salary Level"] = pd.cut(people_df["Salary"], bins, labels=bin_names)
people_df.head()
