==> activity10/15_culimination-02.ipynb <==
# Game of Thrones: Pandas Are Coming

## 1. Import Libraries and Dependencies

# Import necessary libraries and dependencies
import pandas as pd 
from pathlib import Path

## 2. Set File Path to CSV using Pathlib

# Save file path to variable
got_path = Path("./Resources/game_of_thrones.csv")

## 3. Import CSV into Pandas DataFrame

# Read csv with Pandas
got_df = pd.read_csv(got_path)
got_df.head()

## 4. Count Unique Allegiances

# Collect a list of all the unique values in "Allegiances"
got_df["Allegiances"].unique()

## 5. Filter to Only Allegiances and Names

# Create a new table with only the Allegiances and Names
filtered_df = got_df[["Allegiances", "Name"]]
filtered_df.head()

## 6. Groupby Allegiances and Count the number of Members

# Create a GroupBy object based on the "Allegiances" column in the newley created table
got_grp = filtered_df.groupby(["Allegiances"])

got_grp_count = got_grp.count()
got_grp_count.head()

## 7. Rename Column from `Name` to `Members`

# Use the `rename` function to change the `Name` column to `Members`
got_grp_count.columns = ["Members"]
got_grp_count

## 8. Sort House Member Count in Descending Order

# Sort by most deaths and reset index
got_grp_count.sort_values(["Members"], ascending=False, inplace=True)

## 9. Select the 2nd Largest Allegiance via iloc

# Save the death count for the Night's Watch, using `iloc` to filter
night_watch = got_grp_count.iloc[1]
night_watch
