==> activity10/17_culmination-04.ipynb <==
# Gotta Catch em All: Listing the Strongest Pokemon

## 1. Import Libraries and Dependencies

# import libraries and dependencies
import pandas as pd
from pathlib import Path

## 2. Set the File Path to the CSV using Pathlib

# Set the path of the Pokemon csv file to be loaded from the Resources folder.
# Define the path name as pokemon_path
pokemon_path = Path("./Resources/Pokemon.csv")

## 3. Import the CSV into Pandas DataFrame

# Define a variable "pokemon_df" and store the Pokemon.csv into Pandas DataFrame using the pokemon_path
pokemon_df = pd.read_csv(pokemon_path)
# Print out the first 5 rows.
pokemon_df.head()

## 4. View Summary Statistics

# Use the describe function on pokemon_df to review general info on Total percentiles.
pokemon_df.describe()

## 5. Create Bins for Total Ranking

# Create the variable bins to define our bounds for when we cut the dataframe.
# Bins are 0, 330, 450, 515, 600, 700, 800
bins = [0, 330, 450,515, 600, 700, 800]
# Create names for the bins
# group_names to be "Very Weak", "Weak", "Moderate", "Strong", "Very Strong", "Over Powered"
group_names = ["Very Weak", "Weak", "Moderate", "Strong", "Very Strong", "Over Powered"]
pokemon_df

## 6. Label Records According to the Corresponding Bin

# Add a column named "Total Ranking" to pokemon_df and slice the data into the bins
pokemon_df["Total Ranking"] = pd.cut(pokemon_df["Total"], bins, labels=group_names)
# Print out the first 10 rows of pokemon_df.
pokemon_df.head(30)

## 7. Set the Index to the DataFrame

# Set new index to Total Ranking
pokemon_df.set_index("Total Ranking")

## 8. Filter Down to the Strongest Pokemon

# Set the variable overpowered_df to be the pokemon dataframe with Over Powered only.
overpowered_df = pokemon_df.loc[pokemon_df["Total Ranking"] == "Over Powered", :]
overpowered_df
